Extract serde and graphql helpers into document_config submodules
