+++ crates/defra-agent/src/document_config/mod.rs
@@ -5,12 +5,9 @@ mod tool_selection;
 mod serde_helpers;
 mod graphql_fields;
 
-use std::fmt;
-
 use anyhow::{anyhow, Result};
 use defra_node::EmbeddedNode;
-use serde::de::{DeserializeOwned, SeqAccess, Visitor};
-use serde::{Deserialize, Deserializer, Serialize};
+use serde::{Deserialize, Serialize};
 
 use crate::graphql::escape_graphql_string;
 
@@ -64,10 +61,10 @@ pub struct ToolSelectionDocument {
     pub file_tool_root: Option<String>,
     pub enable_bash: Option<bool>,
     pub bash_mode: Option<String>,
-    #[serde(default, deserialize_with = "deserialize_optional_string_vec")]
+    #[serde(default, deserialize_with = "serde_helpers::deserialize_optional_string_vec")]
     pub cli_tool_names: Option<Vec<String>>,
     pub enable_meta_tools: Option<bool>,
-    #[serde(default, deserialize_with = "deserialize_optional_string_vec")]
+    #[serde(default, deserialize_with = "serde_helpers::deserialize_optional_string_vec")]
     pub delegate_to: Option<Vec<String>>,
 }
 
@@ -90,7 +87,7 @@ pub async fn ensure_agent_principal(
     let existing_principal = load_agent_principal(node, agent_did).await?;
     let (default_behavior_id, created_principal) = match existing_principal.as_ref() {
         Some(principal) => {
-            let behavior_id = normalize_optional_string(principal.default_behavior_id.as_deref())
+            let behavior_id = serde_helpers::normalize_optional_string(principal.default_behavior_id.as_deref())
                 .map(ToOwned::to_owned)
                 .unwrap_or_else(|| default_behavior_id_for_agent(agent_did));
             (behavior_id, false)
@@ -117,7 +114,7 @@ pub async fn ensure_agent_principal(
             if existing_principal
                 .as_ref()
                 .and_then(|principal| {
-                    normalize_optional_string(principal.default_behavior_id.as_deref())
+                    serde_helpers::normalize_optional_string(principal.default_behavior_id.as_deref())
                 })
                 .is_some()
             {
@@ -138,8 +135,8 @@ pub async fn ensure_agent_principal(
 
     match existing_principal {
         Some(principal) => {
-            if normalize_optional_string(principal.default_behavior_id.as_deref()).is_none() {
-                let fallback_display_name = default_display_name_for_did(agent_did);
+            if serde_helpers::normalize_optional_string(principal.default_behavior_id.as_deref()).is_none() {
+                let fallback_display_name = serde_helpers::default_display_name_for_did(agent_did);
                 upsert_agent_principal(
                     node,
                     agent_did,
@@ -154,7 +151,7 @@ pub async fn ensure_agent_principal(
             }
         }
         None => {
-            let fallback_display_name = default_display_name_for_did(agent_did);
+            let fallback_display_name = serde_helpers::default_display_name_for_did(agent_did);
             upsert_agent_principal(
                 node,
                 agent_did,
@@ -214,7 +211,7 @@ pub(crate) async fn load_agent_principal_record(
         anyhow::bail!("query AgentPrincipal failed: {:?}", resp.errors);
     }
 
-    Ok(first_row_with_doc_id(resp.data.as_ref(), "AgentPrincipal"))
+    Ok(serde_helpers::first_row_with_doc_id(resp.data.as_ref(), "AgentPrincipal"))
 }
 
 pub(crate) async fn load_agent_principal_by_doc_id(
@@ -244,7 +241,7 @@ pub(crate) async fn load_agent_principal_by_doc_id(
         anyhow::bail!("query AgentPrincipal by _docID failed: {:?}", resp.errors);
     }
 
-    Ok(first_row_with_doc_id(resp.data.as_ref(), "AgentPrincipal"))
+    Ok(serde_helpers::first_row_with_doc_id(resp.data.as_ref(), "AgentPrincipal"))
 }
 
 pub async fn load_agent_behavior(
@@ -289,7 +286,7 @@ pub(crate) async fn load_agent_behavior_record(
         anyhow::bail!("query AgentBehavior failed: {:?}", resp.errors);
     }
 
-    Ok(first_row_with_doc_id(resp.data.as_ref(), "AgentBehavior"))
+    Ok(serde_helpers::first_row_with_doc_id(resp.data.as_ref(), "AgentBehavior"))
 }
 
 pub(crate) async fn load_agent_behavior_by_doc_id(
@@ -325,7 +322,7 @@ pub(crate) async fn load_agent_behavior_by_doc_id(
         anyhow::bail!("query AgentBehavior by _docID failed: {:?}", resp.errors);
     }
 
-    Ok(first_row_with_doc_id(resp.data.as_ref(), "AgentBehavior"))
+    Ok(serde_helpers::first_row_with_doc_id(resp.data.as_ref(), "AgentBehavior"))
 }
 
 pub async fn list_agent_behaviors(
@@ -372,7 +369,7 @@ pub(crate) async fn list_agent_behavior_records(
         anyhow::bail!("list AgentBehavior failed: {:?}", resp.errors);
     }
 
-    Ok(rows_with_doc_id(resp.data.as_ref(), "AgentBehavior"))
+    Ok(serde_helpers::rows_with_doc_id(resp.data.as_ref(), "AgentBehavior"))
 }
 
 pub async fn load_inference_profile(
@@ -413,7 +410,7 @@ pub(crate) async fn load_inference_profile_record(
         anyhow::bail!("query InferenceProfile failed: {:?}", resp.errors);
     }
 
-    Ok(first_row_with_doc_id(
+    Ok(serde_helpers::first_row_with_doc_id(
         resp.data.as_ref(),
         "InferenceProfile",
     ))
@@ -448,7 +445,7 @@ pub(crate) async fn load_inference_profile_by_doc_id(
         anyhow::bail!("query InferenceProfile by _docID failed: {:?}", resp.errors);
     }
 
-    Ok(first_row_with_doc_id(
+    Ok(serde_helpers::first_row_with_doc_id(
         resp.data.as_ref(),
         "InferenceProfile",
     ))
@@ -495,7 +492,7 @@ pub(crate) async fn load_tool_selection_record(
         anyhow::bail!("query ToolSelection failed: {:?}", resp.errors);
     }
 
-    Ok(first_row_with_doc_id(resp.data.as_ref(), "ToolSelection"))
+    Ok(serde_helpers::first_row_with_doc_id(resp.data.as_ref(), "ToolSelection"))
 }
 
 pub(crate) async fn load_tool_selection_by_doc_id(
@@ -530,7 +527,7 @@ pub(crate) async fn load_tool_selection_by_doc_id(
         anyhow::bail!("query ToolSelection by _docID failed: {:?}", resp.errors);
     }
 
-    Ok(first_row_with_doc_id(resp.data.as_ref(), "ToolSelection"))
+    Ok(serde_helpers::first_row_with_doc_id(resp.data.as_ref(), "ToolSelection"))
 }
 
 pub(crate) async fn list_tool_selection_records(
@@ -565,7 +562,7 @@ pub(crate) async fn list_tool_selection_records(
         anyhow::bail!("list ToolSelection failed: {:?}", resp.errors);
     }
 
-    Ok(rows_with_doc_id(resp.data.as_ref(), "ToolSelection"))
+    Ok(serde_helpers::rows_with_doc_id(resp.data.as_ref(), "ToolSelection"))
 }
 
 pub(crate) async fn list_all_tool_selection_records(
@@ -593,7 +590,7 @@ pub(crate) async fn list_all_tool_selection_records(
         anyhow::bail!("list all ToolSelection failed: {:?}", resp.errors);
     }
 
-    Ok(rows_with_doc_id(resp.data.as_ref(), "ToolSelection"))
+    Ok(serde_helpers::rows_with_doc_id(resp.data.as_ref(), "ToolSelection"))
 }
 
 pub(crate) async fn list_inference_profile_records(
@@ -618,7 +615,7 @@ pub(crate) async fn list_inference_profile_records(
         anyhow::bail!("list InferenceProfile failed: {:?}", resp.errors);
     }
 
-    Ok(rows_with_doc_id(resp.data.as_ref(), "InferenceProfile"))
+    Ok(serde_helpers::rows_with_doc_id(resp.data.as_ref(), "InferenceProfile"))
 }
 
 pub async fn upsert_agent_behavior(node: &EmbeddedNode, behavior: &AgentBehavior) -> Result<()> {
@@ -633,21 +630,21 @@ pub async fn upsert_agent_behavior(node: &EmbeddedNode, behavior: &AgentBehavior
     let add_fields = vec![
         Some(format!(r#"behavior_id: "{escaped_behavior_id}""#)),
         Some(format!(r#"agent_did: "{escaped_agent_did}""#)),
-        graphql_string_field("display_name", behavior.display_name.as_deref()),
-        graphql_string_field("system_prompt", behavior.system_prompt.as_deref()),
-        graphql_string_field("backend_id", behavior.backend_id.as_deref()),
-        graphql_string_field("model_name", behavior.model_name.as_deref()),
-        graphql_string_field("tool_selection_id", behavior.tool_selection_id.as_deref()),
-        graphql_string_field(
+        graphql_fields::graphql_string_field("display_name", behavior.display_name.as_deref()),
+        graphql_fields::graphql_string_field("system_prompt", behavior.system_prompt.as_deref()),
+        graphql_fields::graphql_string_field("backend_id", behavior.backend_id.as_deref()),
+        graphql_fields::graphql_string_field("model_name", behavior.model_name.as_deref()),
+        graphql_fields::graphql_string_field("tool_selection_id", behavior.tool_selection_id.as_deref()),
+        graphql_fields::graphql_string_field(
             "inference_profile_id",
             behavior.inference_profile_id.as_deref(),
         ),
-        graphql_string_field(
+        graphql_fields::graphql_string_field(
             "compaction_strategy",
             behavior.compaction_strategy.as_deref(),
         ),
-        graphql_optional_float_field("compaction_threshold", behavior.compaction_threshold),
-        Some(format!("enabled: {}", graphql_bool(behavior.enabled))),
+        graphql_fields::graphql_optional_float_field("compaction_threshold", behavior.compaction_threshold),
+        Some(format!("enabled: {}", graphql_fields::graphql_bool(behavior.enabled))),
         Some(format!(
             r#"created_at: "{}""#,
             escape_graphql_string(created_at.as_str())
@@ -660,21 +657,21 @@ pub async fn upsert_agent_behavior(node: &EmbeddedNode, behavior: &AgentBehavior
 
     let update_fields = vec![
         Some(format!(r#"agent_did: "{escaped_agent_did}""#)),
-        graphql_string_field("display_name", behavior.display_name.as_deref()),
-        graphql_string_field("system_prompt", behavior.system_prompt.as_deref()),
-        graphql_string_field("backend_id", behavior.backend_id.as_deref()),
-        graphql_string_field("model_name", behavior.model_name.as_deref()),
-        graphql_string_field("tool_selection_id", behavior.tool_selection_id.as_deref()),
-        graphql_string_field(
+        graphql_fields::graphql_string_field("display_name", behavior.display_name.as_deref()),
+        graphql_fields::graphql_string_field("system_prompt", behavior.system_prompt.as_deref()),
+        graphql_fields::graphql_string_field("backend_id", behavior.backend_id.as_deref()),
+        graphql_fields::graphql_string_field("model_name", behavior.model_name.as_deref()),
+        graphql_fields::graphql_string_field("tool_selection_id", behavior.tool_selection_id.as_deref()),
+        graphql_fields::graphql_string_field(
             "inference_profile_id",
             behavior.inference_profile_id.as_deref(),
         ),
-        graphql_string_field(
+        graphql_fields::graphql_string_field(
             "compaction_strategy",
             behavior.compaction_strategy.as_deref(),
         ),
-        graphql_optional_float_field("compaction_threshold", behavior.compaction_threshold),
-        Some(format!("enabled: {}", graphql_bool(behavior.enabled))),
+        graphql_fields::graphql_optional_float_field("compaction_threshold", behavior.compaction_threshold),
+        Some(format!("enabled: {}", graphql_fields::graphql_bool(behavior.enabled))),
     ]
     .into_iter()
     .flatten()
@@ -712,18 +709,18 @@ pub async fn upsert_tool_selection(
     let add_fields = vec![
         Some(format!(r#"selection_id: "{escaped_selection_id}""#)),
         Some(format!(r#"agent_did: "{escaped_agent_did}""#)),
-        graphql_string_field("display_name", selection.display_name.as_deref()),
-        graphql_optional_bool_field("enable_file_tools", selection.enable_file_tools),
-        graphql_string_field("file_tools_mode", selection.file_tools_mode.as_deref()),
-        Some(graphql_nullable_string_field(
+        graphql_fields::graphql_string_field("display_name", selection.display_name.as_deref()),
+        graphql_fields::graphql_optional_bool_field("enable_file_tools", selection.enable_file_tools),
+        graphql_fields::graphql_string_field("file_tools_mode", selection.file_tools_mode.as_deref()),
+        Some(graphql_fields::graphql_nullable_string_field(
             "file_tool_root",
             selection.file_tool_root.as_deref(),
         )),
-        graphql_optional_bool_field("enable_bash", selection.enable_bash),
-        graphql_string_field("bash_mode", selection.bash_mode.as_deref()),
-        graphql_string_list_field("cli_tool_names", selection.cli_tool_names.as_deref()),
-        graphql_optional_bool_field("enable_meta_tools", selection.enable_meta_tools),
-        graphql_string_list_field("delegate_to", selection.delegate_to.as_deref()),
+        graphql_fields::graphql_optional_bool_field("enable_bash", selection.enable_bash),
+        graphql_fields::graphql_string_field("bash_mode", selection.bash_mode.as_deref()),
+        graphql_fields::graphql_string_list_field("cli_tool_names", selection.cli_tool_names.as_deref()),
+        graphql_fields::graphql_optional_bool_field("enable_meta_tools", selection.enable_meta_tools),
+        graphql_fields::graphql_string_list_field("delegate_to", selection.delegate_to.as_deref()),
     ]
     .into_iter()
     .flatten()
@@ -732,18 +729,18 @@ pub async fn upsert_tool_selection(
 
     let update_fields = vec![
         Some(format!(r#"agent_did: "{escaped_agent_did}""#)),
-        graphql_string_field("display_name", selection.display_name.as_deref()),
-        graphql_optional_bool_field("enable_file_tools", selection.enable_file_tools),
-        graphql_string_field("file_tools_mode", selection.file_tools_mode.as_deref()),
-        Some(graphql_nullable_string_field(
+        graphql_fields::graphql_string_field("display_name", selection.display_name.as_deref()),
+        graphql_fields::graphql_optional_bool_field("enable_file_tools", selection.enable_file_tools),
+        graphql_fields::graphql_string_field("file_tools_mode", selection.file_tools_mode.as_deref()),
+        Some(graphql_fields::graphql_nullable_string_field(
             "file_tool_root",
             selection.file_tool_root.as_deref(),
         )),
-        graphql_optional_bool_field("enable_bash", selection.enable_bash),
-        graphql_string_field("bash_mode", selection.bash_mode.as_deref()),
-        graphql_string_list_field("cli_tool_names", selection.cli_tool_names.as_deref()),
-        graphql_optional_bool_field("enable_meta_tools", selection.enable_meta_tools),
-        graphql_string_list_field("delegate_to", selection.delegate_to.as_deref()),
+        graphql_fields::graphql_optional_bool_field("enable_bash", selection.enable_bash),
+        graphql_fields::graphql_string_field("bash_mode", selection.bash_mode.as_deref()),
+        graphql_fields::graphql_string_list_field("cli_tool_names", selection.cli_tool_names.as_deref()),
+        graphql_fields::graphql_optional_bool_field("enable_meta_tools", selection.enable_meta_tools),
+        graphql_fields::graphql_string_list_field("delegate_to", selection.delegate_to.as_deref()),
     ]
     .into_iter()
     .flatten()
@@ -779,13 +776,13 @@ pub async fn upsert_inference_profile(
 
     let add_fields = vec![
         Some(format!(r#"profile_id: "{escaped_profile_id}""#)),
-        graphql_string_field("display_name", profile.display_name.as_deref()),
-        graphql_optional_int_field("context_window", profile.context_window),
-        graphql_optional_int_field("max_output_tokens", profile.max_output_tokens),
-        graphql_optional_int_field("max_turns", profile.max_turns),
-        graphql_optional_float_field("temperature", profile.temperature),
-        graphql_optional_int_field("stream_batch_ms", profile.stream_batch_ms),
-        graphql_optional_int_field("deadline_duration_secs", profile.deadline_duration_secs),
+        graphql_fields::graphql_string_field("display_name", profile.display_name.as_deref()),
+        graphql_fields::graphql_optional_int_field("context_window", profile.context_window),
+        graphql_fields::graphql_optional_int_field("max_output_tokens", profile.max_output_tokens),
+        graphql_fields::graphql_optional_int_field("max_turns", profile.max_turns),
+        graphql_fields::graphql_optional_float_field("temperature", profile.temperature),
+        graphql_fields::graphql_optional_int_field("stream_batch_ms", profile.stream_batch_ms),
+        graphql_fields::graphql_optional_int_field("deadline_duration_secs", profile.deadline_duration_secs),
     ]
     .into_iter()
     .flatten()
@@ -793,13 +790,13 @@ pub async fn upsert_inference_profile(
     .join(",\n                    ");
 
     let update_fields = vec![
-        graphql_string_field("display_name", profile.display_name.as_deref()),
-        graphql_optional_int_field("context_window", profile.context_window),
-        graphql_optional_int_field("max_output_tokens", profile.max_output_tokens),
-        graphql_optional_int_field("max_turns", profile.max_turns),
-        graphql_optional_float_field("temperature", profile.temperature),
-        graphql_optional_int_field("stream_batch_ms", profile.stream_batch_ms),
-        graphql_optional_int_field("deadline_duration_secs", profile.deadline_duration_secs),
+        graphql_fields::graphql_string_field("display_name", profile.display_name.as_deref()),
+        graphql_fields::graphql_optional_int_field("context_window", profile.context_window),
+        graphql_fields::graphql_optional_int_field("max_output_tokens", profile.max_output_tokens),
+        graphql_fields::graphql_optional_int_field("max_turns", profile.max_turns),
+        graphql_fields::graphql_optional_float_field("temperature", profile.temperature),
+        graphql_fields::graphql_optional_int_field("stream_batch_ms", profile.stream_batch_ms),
+        graphql_fields::graphql_optional_int_field("deadline_duration_secs", profile.deadline_duration_secs),
     ]
     .into_iter()
     .flatten()
@@ -860,12 +857,12 @@ pub async fn upsert_agent_principal(
     enabled: bool,
 ) -> Result<()> {
     let escaped_agent_did = escape_graphql_string(agent_did);
-    let fallback_display_name = default_display_name_for_did(agent_did);
+    let fallback_display_name = serde_helpers::default_display_name_for_did(agent_did);
     let display_name =
-        normalize_optional_string(display_name).unwrap_or(fallback_display_name.as_str());
+        serde_helpers::normalize_optional_string(display_name).unwrap_or(fallback_display_name.as_str());
     let escaped_display_name = escape_graphql_string(display_name);
     let escaped_default_behavior_id =
-        escape_graphql_string(normalize_optional_string(default_behavior_id).unwrap_or_default());
+        escape_graphql_string(serde_helpers::normalize_optional_string(default_behavior_id).unwrap_or_default());
     let escaped_created_by = escape_graphql_string(agent_did);
     let created_at = chrono::Utc::now().to_rfc3339();
     let mutation = format!(
@@ -894,166 +891,6 @@ pub async fn upsert_agent_principal(
         anyhow::bail!("upsert AgentPrincipal failed: {:?}", resp.errors);
     }
     Ok(())
-}
-
-fn default_display_name_for_did(agent_did: &str) -> String {
-    agent_did
-        .rsplit(':')
-        .next()
-        .filter(|segment| !segment.trim().is_empty())
-        .unwrap_or(agent_did)
-        .to_string()
-}
-
-fn normalize_optional_string(value: Option<&str>) -> Option<&str> {
-    value.and_then(|value| {
-        let trimmed = value.trim();
-        (!trimmed.is_empty()).then_some(trimmed)
-    })
-}
-
-fn deserialize_optional_string_vec<'de, D>(
-    deserializer: D,
-) -> std::result::Result<Option<Vec<String>>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    struct OptionalStringVecVisitor;
-
-    impl<'de> Visitor<'de> for OptionalStringVecVisitor {
-        type Value = Option<Vec<String>>;
-
-        fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-            formatter.write_str("a string list, null, or empty string")
-        }
-
-        fn visit_unit<E>(self) -> std::result::Result<Self::Value, E>
-        where
-            E: serde::de::Error,
-        {
-            Ok(None)
-        }
-
-        fn visit_none<E>(self) -> std::result::Result<Self::Value, E>
-        where
-            E: serde::de::Error,
-        {
-            Ok(None)
-        }
-
-        fn visit_str<E>(self, value: &str) -> std::result::Result<Self::Value, E>
-        where
-            E: serde::de::Error,
-        {
-            if value.trim().is_empty() {
-                Ok(Some(Vec::new()))
-            } else {
-                Ok(Some(vec![value.to_string()]))
-            }
-        }
-
-        fn visit_string<E>(self, value: String) -> std::result::Result<Self::Value, E>
-        where
-            E: serde::de::Error,
-        {
-            self.visit_str(&value)
-        }
-
-        fn visit_seq<A>(self, mut seq: A) -> std::result::Result<Self::Value, A::Error>
-        where
-            A: SeqAccess<'de>,
-        {
-            let mut values = Vec::new();
-            while let Some(value) = seq.next_element::<String>()? {
-                values.push(value);
-            }
-            Ok(Some(values))
-        }
-    }
-
-    deserializer.deserialize_any(OptionalStringVecVisitor)
-}
-
-fn first_row_with_doc_id<T>(data: Option<&serde_json::Value>, field: &str) -> Option<(String, T)>
-where
-    T: DeserializeOwned,
-{
-    rows_with_doc_id(data, field).into_iter().next()
-}
-
-fn rows_with_doc_id<T>(data: Option<&serde_json::Value>, field: &str) -> Vec<(String, T)>
-where
-    T: DeserializeOwned,
-{
-    data.and_then(|data| data.get(field))
-        .and_then(|value| value.as_array())
-        .map(|rows| {
-            rows.iter()
-                .filter_map(|row| {
-                    let doc_id = row.get("_docID")?.as_str()?.to_string();
-                    let parsed = match serde_json::from_value(row.clone()) {
-                        Ok(parsed) => parsed,
-                        Err(error) => {
-                            tracing::warn!(
-                                field = field,
-                                doc_id = %doc_id,
-                                error = %error,
-                                "failed to deserialize document row"
-                            );
-                            return None;
-                        }
-                    };
-                    Some((doc_id, parsed))
-                })
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-fn graphql_string_field(name: &str, value: Option<&str>) -> Option<String> {
-    Some(format!(
-        r#"{name}: "{}""#,
-        escape_graphql_string(normalize_optional_string(value).unwrap_or_default())
-    ))
-}
-
-fn graphql_nullable_string_field(name: &str, value: Option<&str>) -> String {
-    match normalize_optional_string(value) {
-        Some(value) => format!(r#"{name}: "{}""#, escape_graphql_string(value)),
-        None => format!("{name}: null"),
-    }
-}
-
-fn graphql_optional_int_field(name: &str, value: Option<i64>) -> Option<String> {
-    value.map(|value| format!("{name}: {value}"))
-}
-
-fn graphql_optional_float_field(name: &str, value: Option<f64>) -> Option<String> {
-    value.map(|value| format!("{name}: {value}"))
-}
-
-fn graphql_optional_bool_field(name: &str, value: Option<bool>) -> Option<String> {
-    value.map(|value| format!("{name}: {}", graphql_bool(value)))
-}
-
-fn graphql_string_list_field(name: &str, value: Option<&[String]>) -> Option<String> {
-    let values = value?;
-    Some(format!(
-        "{name}: [{}]",
-        values
-            .iter()
-            .map(|value| format!(r#""{}""#, escape_graphql_string(value)))
-            .collect::<Vec<_>>()
-            .join(", ")
-    ))
-}
-
-fn graphql_bool(value: bool) -> &'static str {
-    if value {
-        "true"
-    } else {
-        "false"
-    }
 }
 
 #[cfg(test)]
